chore(folio): extract DocxEditor types to DocxEditor.props
The 4,247-line `DocxEditor.tsx` becomes harder to read once you can't
see the prop and ref types without scrolling past 500 lines. Move the
type definitions into a sibling `DocxEditor.props.ts` and re-export the
public types (`DocxEditorProps`, `DocxEditorRef`, `EditorMode`) from
`DocxEditor.tsx` so callers can keep their existing import paths.

Pure mechanical move plus a few small clean-ups co-located with the
types being moved:

- Promote the inline `import(...).AnonymizationMatch` reference to a
  proper top-level import.
- Extract the inline `pmImageContext` and `activeTrackedChange` shapes
  into named `ImageContextInfo` and `ActiveTrackedChangeInfo` aliases.
- Consolidate the duplicated `scrollToBlock` doc-comment into one.

No runtime behavior change.
